+++ src/main/java/es/deusto/sd/strava/entity/Challenge.java
@@ -27,7 +27,7 @@ public class Challenge {
     private Float targetDistance; // km, optional
     
 	@Column(name = "target_time")
-    private Float targetTime; // optional
+    private Integer targetTime; // optional
     
 	@Column(name = "start_date", nullable = false)
     private LocalDate startDate;
@@ -44,7 +44,7 @@ public class Challenge {
 
     public Challenge() {}
 
-    public Challenge(String name, String sport, Float targetDistance, Float targetTime,
+    public Challenge(String name, String sport, Float targetDistance, Integer targetTime,
                      LocalDate startDate, LocalDate endDate, LocalDateTime createdAt, Long creatorId) {
         this.name = name;
         this.sport = sport;
@@ -84,11 +84,11 @@ public class Challenge {
 		this.targetDistance = targetDistance;
 	}
 
-	public Float getTargetTime() {
+	public Integer getTargetTime() {
 		return targetTime;
 	}
 
-	public void setTargetTime(Float targetTime) {
+	public void setTargetTime(Integer targetTime) {
 		this.targetTime = targetTime;
 	}
